+++ src/pages/CoursePage/CoursePage.tsx
@@ -1,6 +1,33 @@
-import Header from "../../components/Common/Header/Header";
-
+/* import Header from "../components/Common/Header"; */
+import CourseCard from "../../components/Common/CourseCard/CourseCard";
+import Exercises from "../../components/OtherComponents/Exercises/Exercises";
+import MyProgressModal from "../../components/OtherModals/MyProgressModal/MyProgressModal";
+import WorkoutModal from "../../components/OtherModals/WorkoutModal/WorkoutModal";
 
 export default function CoursePage() {
-  return <Header />;
+  const objCard = [
+    { id: 1, name: "Йога", progress: 40 },
+    { id: 2, name: "Стретчинг", progress: 0 },
+    { id: 3, name: "Зумба", progress: 100 },
+  ];
+  return (
+    <>
+      <div className="w-cardBlockWidth h-cardBlockHeight flex flex-col justify-center">
+        <p className="flex flex-row text-3xl text-sky-50 mb-10">Мои курсы</p>
+        <div className="flex justify-between">
+          {objCard.map((el) => (
+            <CourseCard
+              id={el.id}
+              key={el.id}
+              cursName={el.name}
+              progress={el.progress}
+            />
+          ))}
+        </div>
+      </div>
+      <Exercises />
+      <WorkoutModal />
+      <MyProgressModal />
+    </>
+  );
 }
